`cv()` function moved to `_common_new` module
