+++ extras/_common_new.py
@@ -270,3 +270,15 @@ if __name__ == '__main__':
                                 ELECTRODES.Y,
                                 ELECTRODES.Z)
     assert np.abs((DF.OLD - DF.NEW) / DF.OLD).max() < 1e-10
+
+
+def cv(reconstructor, measured, regularization_parameters):
+    errors = []
+
+    for regularization_parameter in regularization_parameters:
+        logger.info('cv(): error estimation for regularization parameter: {:g}'.format(regularization_parameter))
+        ERR = np.array(reconstructor.leave_one_out_errors(measured,
+                                                          regularization_parameter))
+        errors.append(np.sqrt((ERR**2).mean()))
+
+    return errors
